refactor: LoadingState 컴포넌트 공용 승격
