+++ src/shared/components/loadingState/index.ts
@@ -1,0 +1,2 @@
+export { default } from './LoadingState';
+export type { LoadingStateProps } from './LoadingState';
